requirements.txt
ipdb~=0.13.9
torchmetrics~=0.7.0
transformers~=4.16.2
numpy~=1.21.2
tqdm~=4.62.3
matplotlib~=3.5.1
timm~=0.5.4
sacred~=0.8.2
Pillow~=8.4.0
pandas~=1.4.0
scikit-learn~=1.0.2
fastai~=2.5.3
ipykernel~=6.7.0
pytorch-lightning~=1.5.9